Split message types
