+++ packages/types/src/index.ts
@@ -11,6 +11,7 @@ export * from './discord/guildScheduledEvents.js'
 export * from './discord/guildTemplates.js'
 export * from './discord/interactions.js'
 export * from './discord/invites.js'
+export * from './discord/messages.js'
 export * from './discord/reference.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
